refactor: move details builder into its own file
